Error handling in JniInputStream.h reading and add documentation in native side.

+++ libjxl/src/main/cpp/includes/InputSource.h
@@ -5,9 +5,23 @@
 #ifndef JXLVIEWER_INPUTSOURCE_H
 #define JXLVIEWER_INPUTSOURCE_H
 
+/**
+ * @class InputSource
+ * @brief Abstraction over an input source.
+ *
+ * This represent an abstraction over a source, for example an Java InputStream or a file descriptor.
+ */
 class InputSource {
 public:
-    virtual size_t read(uint8_t *buffer, size_t size) = 0;
+    /**
+     * @brief Reads data from the source into the provided buffer.
+     * @param buffer The buffer to read data into.
+     * @param size The number of bytes to read.
+     * @return The total number of bytes read, -1 if EOF or INT32_MIN if error with JNI error set correctly.
+     *
+     * This method reads data from the source into the specified buffer.
+     */
+    virtual int32_t read(uint8_t *buffer, size_t size) = 0;
 };
 
 #endif //JXLVIEWER_INPUTSOURCE_H
